데이터타입 Integer -> Long으로 바꿨으므로 Repository도 맞게 수정함

+++ src/main/java/net/softbell/bsh/domain/repository/MemberGroupRepo.java
@@ -10,7 +10,7 @@ import net.softbell.bsh.domain.entity.MemberGroup;
  * @Description : 회원 그룹 리포지토리 인터페이스
  */
 @Repository
-public interface MemberGroupRepo extends JpaRepository<MemberGroup, Integer>
+public interface MemberGroupRepo extends JpaRepository<MemberGroup, Long>
 {
 	
 }
